Fix various plotting issues
A number of plotting problems are fixed by first mapping coordinates
from lon/lat to projection coordinates and then plotting, rather than
using the latlon argument to the basemap plot routine.

Added some output so users are aware of what is being plotted and
get a sense of how long it will take.