Add a script to mask out (difference) features
One or more features are supplied with the -f flag and one or more
masking features are supplied with -m.  The result is the first
set of features with the intersection with the second set of
features removed (i.e. the difference as defined by the shapely
library).  The results are stored in features.geojson.