Add a driver script for merging Antarctic basins
The script combines all landice regions with names
staring with Antarctica_IMBIE (because there is not
a convenient tag for this) and splits any basins
that cross the prime meridian into east and west
parts.  Optionally (with --plot flag), the basins
are plotted using the southpole map projection.

Also, renames setup_basins.py to setup_ocean_basins.py
avoid confusion.