Add simplify_features script
Uses shapely's simplify command to remove redundancies in features.
The tolerance flag can be used to allow more simplification by
allowing features to move within a tolerance.  This my be appropriate,
for example, if the shape was adapted from a contour of a dataset on
a grid (e.g. the Bedmap2 land coverage masks).