Add a script that subdivides transects
This merge adds a script, subdivide_transects.py, that is used
to subivide transects (LineStrings) longer than a maximum length
(in degrees) into an integer number of subsegments. This is useful
because some tools (e.g. the MPAS MaskCreator) don't necessarily
follow the desired path between points that are far apart.