Add "constituents" property to combined feature
When multiple features are combined into a single feature, the
names of the constituent features are retained in the
"constituents" property of the new feature.

Add an error message when the combined feature is of type
"GeometryCollection", indicating that multipe feature types
(regions, points and transects) have likely been combined
by mistake.