Change plotting to use cartopy add_geometries
This merge used cartopy support for shapely geometries to perform
plotting rather than creating plotting objects ourselves.

Also, plotting properties have been improved so that transects
are not plotted as filled shapes.

The bounds of the plot are computed based on the combined bounding
box of all plotted objects (other than the coastlines).