Remove translation by 360 degrees
This merge removes a command to translate the western portin of
a feature by 360 degrees after splitting along the prime meridian.
Instead, longitudes should remain between -180 and 180 degrees.

Updated the docstrings for both fix_features*.py scripts to reflect
the fact that the output file may be specified by the -o flag.