Add 'seed_point' tag to all ocean points
This tag is used to find seed points for the "flood fill" operation
used to make sure the ocean is a single, connected body when setting
up MPAS-Ocean `global_ocean` test cases.

Adding this tag will make it possible to add ocean points that are
not necessarily in the ocean domain for all model configurations.
For example, points below ice shelves would not be in the ocean
domain for setups without ice-shelf cavities.