Rename Ross East --> Western Ross...
...and similarly Ross West --> Eastern Ross.  Rignot et al. (2013)
named these parts of the shelf based on which hemisphere or ice
sheet they're in, but this is very confusing given that the Ross
crosses the antimeridian and thus its eastern side is in West
Antarctica and its western side in East Antarctica.  To be true
to the local cardinal directions, we're renaming the halves in
the more standard way.